Agrega modal cuando subes de nivel
