+++ src/app/(application)/new/page.tsx
@@ -3,6 +3,7 @@ import { redirect } from "next/navigation"
 import DeedCreate from "@/components/deed-create"
 import NavBar from "@/components/layout/nav-bar"
 import { authOptions } from "@/lib/auth"
+import prisma from "@/lib/prisma"
 import { getCurrentUser } from "@/lib/session"
 
 export default async function NewPage() {
@@ -12,10 +13,20 @@ export default async function NewPage() {
     redirect(authOptions.pages?.signIn || "/sign-in")
   }
 
+  const data = await prisma.user.findUnique({
+    where: {
+      id: user.id
+    }
+  })
+
+  if (!data) {
+    return
+  }
+
   return (
     <>
       <NavBar title="Agregar" />
-      <DeedCreate userId={user.id} />
+      <DeedCreate userId={user.id} levelPoints={data.levelPoints} />
     </>
   )
 }
